requirements.txt
asgiref==3.3.1
astroid==2.4.2
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.12.5
chardet==3.0.4
click==7.1.2
cycler==0.10.0
Django==3.1.5
django-appconf==1.0.4
django-compressor==2.4
django-sass-processor==0.8.2
docopt==0.6.2
et-xmlfile==1.0.1
freeze==3.0
idna==2.10
isort==5.6.4
jdcal==1.4.1
joblib==1.0.0
kiwisolver==1.3.1
lazy-object-proxy==1.4.3
libsass==0.20.1
lxml==4.6.2
matplotlib==3.4.1
mccabe==0.6.1
nltk==3.5
numpy==1.19.4
openpyxl==3.0.5
pandas==1.1.4
Pillow==8.2.0
psycopg2==2.8.6
pylint==2.6.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.4
rcssmin==1.0.6
regex==2020.11.13
requests==2.25.0
rjsmin==1.1.0
scikit-learn==0.24.1
scipy==1.6.3
selenium==3.141.0
six==1.15.0
sklearn==0.0
soupsieve==2.1
sqlparse==0.4.1
threadpoolctl==2.1.0
toml==0.10.2
torch==1.7.1
tqdm==4.54.1
typing-extensions==3.7.4.3
urllib3==1.26.2
uuid==1.30
wrapt==1.12.1
xlrd==1.2.0
